some updates

+++ components/app/deleteButton.tsx
@@ -4,21 +4,36 @@ import { deleteNoteAction } from "@/app/actions";
 import { Id } from "@/convex/_generated/dataModel";
 import { toast } from "sonner";
 import { useRouter } from "next/navigation";
+import { AlertDialog, AlertDialogContent, AlertDialogDescription, AlertDialogFooter, AlertDialogHeader, AlertDialogTitle, AlertDialogTrigger, AlertDialogCancel, AlertDialogAction } from "../ui/alert-dialog";
 export function DeleteButton({id}: {id: Id<"notes">}){
     const router = useRouter()
     return(
-         
-            <Button onClick={async() =>
-                {
-                const result = await deleteNoteAction(id);
-                if (result.error){
-                    toast.error(`${result.error}`)
-                }
-                else{
-                    toast.success(result.success)
-                    router.push("/")
-                }
-                }}
-            className="ml-auto" variant="destructive">Delete</Button>
+        <AlertDialog>
+            <AlertDialogTrigger asChild>
+                <Button variant="destructive">Delete</Button>
+            </AlertDialogTrigger>
+            <AlertDialogContent>
+                <AlertDialogHeader>
+                    <AlertDialogTitle>Are you absolutely sure?</AlertDialogTitle>
+                    <AlertDialogDescription>
+                        This action cannot be undone. This will permanently delete your note and remove your data from our servers.
+                    </AlertDialogDescription>
+                </AlertDialogHeader>
+                <AlertDialogFooter>
+                    <AlertDialogCancel>Cancel</AlertDialogCancel>
+                    <AlertDialogAction onClick={async() =>
+                    {
+                    const result = await deleteNoteAction(id);
+                    if (result.error){
+                        toast.error(`${result.error}`)
+                    }
+                    else{
+                        toast.success(result.success)
+                        router.push("/")
+                    }
+                    }} variant="destructive">Continue</AlertDialogAction>
+                </AlertDialogFooter>
+            </AlertDialogContent>
+        </AlertDialog>
     )
 }
